getion d'erreur

+++ demo/demo/src/main/java/com/example/demo/service/UserService.java
@@ -3,6 +3,7 @@ package com.example.demo.service;
 import com.example.demo.entity.User;
 import com.example.demo.repository.UserRepository;
 import org.springframework.stereotype.Service;
+import com.example.demo.exception.ResourceNotFoundException;
 
 import java.util.List;
 
@@ -23,7 +24,25 @@ public class UserService {
         return userRepository.findAll();
     }
 
+    public User getById(Long id) {
+        return userRepository.findById(id)
+            .orElseThrow(() -> new ResourceNotFoundException("Utilisateur non trouvé"));
+    }
+
+    public User update(Long id, User user) {
+        User existingUser = userRepository.findById(id)
+            .orElseThrow(() -> new ResourceNotFoundException("Utilisateur non trouvé"));
+        
+        existingUser.setName(user.getName());
+        existingUser.setAge(user.getAge());
+
+        return userRepository.save(existingUser);
+    }
+
     public void deleteUserById(Long id) {
+        if(!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Utilisateur non trouvé");
+        }
         userRepository.deleteById(id);
     }
 }
